Renamed the index.js files that are part of the single cause component to the actual name of the component.
