+++ src/screens/SingleCause/components/ProgressBar/ProgressBar.js
